Version-specific compatibility
I've copied over some version compat infrastructure from coq-rewriter,
while improving it and documenting it a bit more.  The basic idea is
that we parse the version of Coq, record it in .coq-version (only when
the version has changed from the last build, though!), tell Coq that all
build outputs depend on .coq-version (this is why it's important that
.coq-version is only changed when there's actually a change in the
version of Coq), and then say that Int63Compat.v is generated by copying
over the relevant version-specific file.